refactor(automations): createAutomationsEngine was 1,980 lines; it is 13
`createAutomationsEngine` held every door and every helper of 07 §1 in one
closure: 1,980 lines inside a 2,499-line file, with the row schemas and the
doors that read them ~2,000 lines apart. The zone's frozen done-means is "no
touched function over 120 lines".

Decomposed in the shape #963/#980 established for `createApps`: a module per
concern, each taking a `Pick` of one shared closure type
(`AutomationsEngineContext`) and returning its slice, wired in dependency order
by `createEngineContext`. `engine.ts` is now the assembler and nothing else.

Pure movement — 97.0% of the removed lines reappear byte-identical, and the
72 that do not are `const X` → `export const X` re-declarations plus import
and return-object lines. No stored shape, no query path and no persisted
sentence changed; the package root's exports are untouched, so nothing outside
`packages/automations/src` moved and no test file changed.
